requirements.txt
asgiref==3.4.1
Babel==2.9.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.8
django-crispy-forms==1.13.0
django-phonenumber-field==5.2.0
idna==3.2
phonenumbers==8.12.34
Pillow==8.3.2
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
stripe==2.60.0
urllib3==1.26.7
